[r] Create a helper that store functions used to build strings
